+++ packages/zendesk-client/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "zendesk-client",
+	"description": "Manages the connection to the Zendesk API",
+	"author": "Automattic Inc.",
+	"main": "dist/cjs/index.js",
+	"module": "dist/esm/index.js",
+	"types": "dist/types",
+	"homepage": "https://github.com/Automattic/wp-calypso",
+	"version": "1.0.0",
+	"sideEffects": false,
+	"scripts": {
+		"build": "tsc --build ./tsconfig.json ./tsconfig-cjs.json",
+		"clean": "tsc --build ./tsconfig.json ./tsconfig-cjs.json --clean && rm -rf dist",
+		"prepare": "yarn run build",
+		"prepack": "yarn run clean && yarn run build",
+		"watch": "tsc --build ./tsconfig.json --watch"
+	},
+	"dependencies": {
+		"@tanstack/react-query": "^5.48.0",
+		"@wordpress/api-fetch": "^7.2.0",
+		"@wordpress/element": "^6.2.0",
+		"@wordpress/url": "^4.2.0"
+	},
+	"devDependencies": {
+		"@automattic/calypso-typescript-config": "workspace:^",
+		"typescript": "^5.3.3",
+		"wpcom-proxy-request": "workspace:^"
+	}
+}
