+++ packages/zendesk-client/src/constants.ts
@@ -1,0 +1,2 @@
+export const ZENDESK_SOURCE_URL_TICKET_FIELD_ID = 23752099174548;
+export const ZENDESK_SCRIPT_ID = 'ze-snippet';
