+++ packages/zendesk-client/src/use-can-connect-to-zendesk-messaging.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-export function useCanConnectToZendesk( enabled = true ) {
+export function useCanConnectToZendeskMessaging( enabled = true ) {
 	return useQuery( {
 		queryKey: [ 'canConnectToZendesk' ],
 		queryFn: async () => {
