+++ packages/zendesk-client/src/use-update-zendesk-user-fields.ts
@@ -1,18 +1,17 @@
+/**
+ * External dependencies
+ */
 import { useMutation } from '@tanstack/react-query';
 import apiFetch, { APIFetchOptions } from '@wordpress/api-fetch';
 import wpcomRequest, { canAccessWpcomApis } from 'wpcom-proxy-request';
+/**
+ * Internal dependencies
+ */
+import type { UserFields } from './types';
 
-type ZendeskUserFields = {
-	messaging_ai_chat_id?: string;
-	messaging_initial_message?: string;
-	messaging_plan?: string;
-	messaging_source?: string;
-	messaging_url?: string;
-};
-
-export function useUpdateZendeskUserFieldsMutation() {
+export function useUpdateZendeskUserFields() {
 	return useMutation( {
-		mutationFn: ( userFields: ZendeskUserFields ) => {
+		mutationFn: ( userFields: UserFields ) => {
 			return canAccessWpcomApis()
 				? wpcomRequest( {
 						path: 'help/zendesk/update-user-fields',
