+++ packages/zendesk-client/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "@automattic/calypso-typescript-config/ts-package.json",
+	"compilerOptions": {
+		"outDir": "dist/esm",
+		"declarationDir": "dist/types",
+		"rootDir": "src"
+	},
+	"include": [ "src", "types.d.ts" ],
+	"exclude": [ "**/test/*" ]
+}
