+++ packages/zendesk-client/types.d.ts
@@ -1,0 +1,9 @@
+interface Window {
+	zE?: (
+		action: string,
+		value: string,
+		handler?:
+			| ( ( callback: ( data: string | number ) => void ) => void )
+			| { id: number; value: string }[]
+	) => void;
+}
